Bitcoin Legacy Gift Coin
The Bitcoin Legacy Gift Coin is a custom designed two-piece coin engineered to
securely and safely gift a Bitcoin cold wallet in a tangible form, reminiscent
of gifting gold.

This repo contains the 3d models which consists of two coins that embed a
bitcoin seed phrase.

The first coin is a generic "Bitcoin" coin, containing a bitcoin logo on the
front and the embedded word "bitcoin" on the reverse.

The second coin contains a unique wallet seed phrase, and the reverse displays
the denomination that the wallet contains (such as "1 Bitcoin" or "1,000,000
Satoshis").

The coins are designed to be joined together, keeping the seed phrase secure
from casual viewing.

An actual production run should be done from a dedicated offline system that is
securely erased after the run.

Signed-off-by: Dan Rue <drue@therub.org>